+++ CHANGELOG.md
@@ -79,6 +79,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   only traverses the first parent of each commit (similar to Git's
   `--first-parent` option).
 
+* New `signing.backends.ssh.revocation-list` config for specifying a list of revoked
+  public keys for commit signature verification.
+
 ### Fixed bugs
 
 * Fixed an error in `jj util gc` caused by the empty blob being missing from